Adicionei o atributo Tipo a classe produto, corrigi também os scans da main, além disso realizei a correção onde era necesário para que o novo atributo não desse conflito ou quebrasse nada do código.
